+++ test/unit/fuzz-tools.test.js
@@ -382,6 +382,7 @@ import * as laugeHansenV402 from '../../lib/lauge-hansen-v402.js';
 import * as berndtHartyV403 from '../../lib/berndt-harty-v403.js';
 import * as reganMorreyV404 from '../../lib/regan-morrey-v404.js';
 import * as savaryMillerV405 from '../../lib/savary-miller-v405.js';
+import * as leFortV406 from '../../lib/le-fort-v406.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -709,6 +710,7 @@ const MODULES = {
   'berndt-harty-v403.js': berndtHartyV403,
   'regan-morrey-v404.js': reganMorreyV404,
   'savary-miller-v405.js': savaryMillerV405,
+  'le-fort-v406.js': leFortV406,
   'tb-testing.js': tbTesting,
 };